BL Outline, General changes
Added a file that described the general process in Black Litterman and
outlines the methodology required; no specifics yet

It appears that IFP-A.TO/IFP.TO is confused and doesn't know which one
it wants to be and needs to be changed whenever it feels like changing.
I can't yet think of a way to verify how to do this atm. In future, will
likely just have to make a script that flat out downloads the entire TSX
list and creats the assets array much like we did, except autonomously
(likely pretty big run time).

Arbitrary start date change, negligible: Perhaps will be important if we
run any of the other programs

Changed maxiter in sMVO as I was supposed to from a few commits ago