Small changes
Found a missing inverse in the E[R] equation

Changed a lot of small logistic things that should be relatively easy to
spot, like outputing the risk aversion coefficient

Tried outputting the weights via the inverse formula in the paper

I tried the actual formula for BL_omega, didn't really yield very good
results BUT

I figured out that if you change the Q matrix at all, the weights change
completely to minimize the variance. As long as the function can meet
the minimum required return, the optimal weight distribution is only
dependent on the covariance matrix. I feel like we should change the
inequality constraint to an equality constraint for the desired return
in order to try and mitigate this